requirements.txt
flask==2.3.3
flask-socketio==5.3.5
flask-bcrypt==1.0.1
pymongo==4.5.0
python-dotenv==1.0.0
eventlet==0.33.3  # For production Socket.IO